calendar functionality done
Earlier, the day-card components were just thrown in, at random, with just 6 cards per row. Now, they are arranged like a calendar, with 7 cards per row and filler cards to complete the week, at the start and end of the month, if necessary. Made changes to the ngrx months reducer to now get weeks as a multidimensional array instead of the days in a month number value.